Move activity bar together with corresponding nav items to subdirectory
