+++ client/src/components/ActivityBar/Items/UploadButton.vue
