+++ client/src/components/Upload/index.js
@@ -2,4 +2,3 @@
  * External entry point for Upload components (currently only the modal).
  */
 export { default as UploadModal } from "./UploadModal";
-export { default as UploadButton } from "./UploadButton";
