+++ client/src/entry/analysis/modules/Analysis.vue
@@ -3,7 +3,7 @@ import { getGalaxyInstance } from "app";
 import CenterFrame from "./CenterFrame";
 import { useUserStore } from "@/stores/userStore";
 import HistoryIndex from "@/components/History/Index.vue";
-import ActivityBar from "@/components/Masthead/ActivityBar.vue";
+import ActivityBar from "@/components/ActivityBar/ActivityBar.vue";
 import { WindowManager } from "@/layout/window-manager";
 import { useRoute, useRouter } from "vue-router/composables";
 import { computed, ref, onMounted, onUnmounted } from "vue";
